move legacy code to the legacy directories
