+++ cpp/include/cugraph/algorithms.hpp
@@ -21,7 +21,7 @@
 #include <cugraph/graph.hpp>
 #include <cugraph/graph_view.hpp>
 
-#include <cugraph/internals.hpp>
+#include <cugraph/legacy/internals.hpp>
 #include <cugraph/legacy/graph.hpp>
 
 #ifndef NO_CUGRAPH_OPS
