+++ cpp/include/cugraph/legacy/internals.hpp
