+++ cpp/src/components/legacy/connectivity.cu
