+++ cpp/src/components/legacy/scc_matrix.cuh
