+++ cpp/src/components/legacy/utils.h
