+++ cpp/src/components/legacy/weak_cc.cuh
