+++ cpp/src/layout/legacy/barnes_hut.cuh
@@ -20,11 +20,11 @@
 #include "fa2_kernels.cuh"
 #include "utils.hpp"
 
-#include <converters/COOtoCSR.cuh>
+#include <converters/legacy/COOtoCSR.cuh>
 #include <utilities/graph_utils.cuh>
 
 #include <cugraph/detail/utility_wrappers.hpp>
-#include <cugraph/internals.hpp>
+#include <cugraph/legacy/internals.hpp>
 #include <cugraph/legacy/graph.hpp>
 #include <cugraph/utilities/error.hpp>
