+++ cpp/src/layout/legacy/exact_repulsion.cuh
