+++ cpp/src/layout/legacy/utils.hpp
