+++ cpp/tests/components/con_comp_test.cu
@@ -20,7 +20,7 @@
 
 #include <rmm/device_vector.hpp>
 
-#include <converters/COOtoCSR.cuh>
+#include <converters/legacy/COOtoCSR.cuh>
 #include <cugraph/algorithms.hpp>
 #include <cugraph/legacy/graph.hpp>
